Add kbd67 mkii rgb v3
