+++ keyboards/kbdfans/kbd67/mkiirgb/v3/readme.md
@@ -1,0 +1,13 @@
+# KBD67 Mk.II RGB V3 (Hotswap)
+
+A hotswap 65% keyboard with USB C, inswitch RGB, and ESD protection.
+
+* Keyboard Maintainer: [moyi4681](https://github.com/moyi4681)
+* Hardware Supported: KBD67 Mk.II RGB V3
+* Hardware Availability: [kbdfans](https://kbdfans.myshopify.com/)
+
+Make example for this keyboard (after setting up your build environment):
+
+    make kbdfans/kbd67/mkiirgb/v3:default
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
